fixed syntax errors in SingleProduct and AllProducts components
