+++ client/components/index.js
@@ -6,3 +6,5 @@
 export {default as Main} from './main'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
+export {default as AllProducts} from './AllProducts'
+export {default as SingleProduct} from './SingleProduct'
