+++ client/routes.js
@@ -4,10 +4,8 @@ import {Router} from 'react-router'
 import {Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
 import history from './history'
-import {Main, Login, Signup, UserHome} from './components'
+import {Main, Login, Signup, UserHome, AllProducts, SingleProduct} from './components'
 import {me, fetchProducts} from './store'
-import AllProducts from './AllProducts'
-import SingleProduct from './SingleProducts'
 
 
 /**
@@ -15,8 +13,7 @@ import SingleProduct from './SingleProducts'
  */
 class Routes extends Component {
   componentDidMount () {
-    this.props.loadInitialData()
-    store.dispatch(fetchProducts());
+    this.props.loadInitialData();
   }
 
   render () {
@@ -29,7 +26,7 @@ class Routes extends Component {
             {/* Routes placed here are available to all visitors */}
             <Route path="/login" component={Login} />
             <Route path="/signup" component={Signup} />
-            <Route path="/products" component = {AllProducts}/>
+            <Route path="/allProducts" component = {AllProducts}/>
             {
               isLoggedIn &&
                 <Switch>
