+++ client/store/index.js
@@ -3,9 +3,9 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import user from './user'
-import products from './store'
+import products from './products'
 
-const reducer = combineReducers({user,products})
+const reducer = combineReducers({user, products})
 const middleware = composeWithDevTools(applyMiddleware(
   thunkMiddleware,
   createLogger({collapsed: true})
